MindCare-Ai-main/requirements.txt
# MindCare AI - Production Dependencies
# Python 3.10+

Flask>=3.0.0
Flask-CORS>=4.0.0
mysql-connector-python>=8.2.0
